requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
langchain==0.0.312
langchain-google-genai==0.1.1
Pillow
pytesseract==0.3.10
pdf2image==1.16.3
fpdf==1.7.2
pydantic==2.4.2
python-dotenv==1.0.0 